% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_data}
\alias{get_data}
\title{Download a table and use optional sql filters}
\usage{
get_data(data_id, sql = NA)
}
\arguments{
\item{data_id}{data id}

\item{sql}{for filtering data NA if you want the whole data}
}
\value{
tibble
}
\description{
Download a table and use optional sql filters
}
