% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan.R
\name{list_ckan}
\alias{list_ckan}
\title{List published data in avoindata.fi CKAN}
\usage{
list_ckan(org = "Kansaneläkelaitos", n = 100)
}
\arguments{
\item{org}{organisation to get listning from}

\item{n}{number of hits}
}
\value{
data.frame
}
\description{
List published data in avoindata.fi CKAN
}
