% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan.R
\name{pick_a_pkg}
\alias{pick_a_pkg}
\title{Get details from a package}
\usage{
pick_a_pkg(input_dataset = "a11f9ec3-669a-4242-bd30-f133a334901b")
}
\arguments{
\item{input_dataset}{}
}
\value{
list
}
\description{
Get details from a package
}
