% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_data_csv}
\alias{get_data_csv}
\title{Download the whole data set using csv-files}
\usage{
get_data_csv(data_id)
}
\arguments{
\item{data_id}{data id}
}
\value{
tibble
}
\description{
Download the whole data set using csv-files
}
