% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckan.R
\name{pick_a_pkg}
\alias{pick_a_pkg}
\title{Get details from a package}
\usage{
pick_a_pkg(input_dataset = "8df7939b-6c0e-4587-9aa1-95c076e7b0a3")
}
\arguments{
\item{input_dataset}{}
}
\value{
list
}
\description{
Get details from a package
}
